Create facialtrack.py
This script handles the facial tracking and picture taking, in which every new picture is taken and added to a locally hosted web server. Also once every picture is taken the script triggers a secondary socket-client process between the facial recognition computer to check for updates on the web server.